// src/services/gql/generated/Login.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: Login
// ====================================================

export interface Login_login {
  __typename: "User";
  id: string;
}

export interface Login {
  login: Login_login | null;
}

export interface LoginVariables {
  usernameOrEmail: string;
  password: string;
}